+++ src/main/java/com/planbase/pdf/lm2/attributes/Orientation.kt
@@ -1,0 +1,6 @@
+package com.planbase.pdf.lm2.attributes
+
+enum class Orientation {
+    PORTRAIT,
+    LANDSCAPE
+}